Converting to doubles

+++ src/edu/umass/cs390cg/atmosphere/Camera.java
@@ -2,66 +2,66 @@ package edu.umass.cs390cg.atmosphere;
 
 import edu.umass.cs390cg.atmosphere.geom.Ray;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 public class Camera {
-  public Vector3f eye, up, at;
-  public float fovy, aspect_ratio;
-  public Vector3f center, corner,across;
+  public Vector3d eye, up, at;
+  public double fovy, aspect_ratio;
+  public Vector3d center, corner,across;
 
-  public Camera(Vector3f eye,
-                Vector3f at,
-                Vector3f up,
-                float fovy,
-                float ar)	{
+  public Camera(Vector3d eye,
+                Vector3d at,
+                Vector3d up,
+                double fovy,
+                double ar)	{
 
     // we will use a default camera
-    this.eye = new Vector3f(eye);
-    this.up = new Vector3f(up);
-    this.at = new Vector3f(at);
+    this.eye = new Vector3d(eye);
+    this.up = new Vector3d(up);
+    this.at = new Vector3d(at);
 
     this.fovy = fovy;
     aspect_ratio = ar;
 
 		/* Code for initializing camera
 		 * Compute the four corner points of the camera's image plane */
-    float dist = 1.f;
-    float top = dist * (float)Math.tan(this.fovy * Math.PI / 360.f);
-    float bottom = -top;
-    float right = aspect_ratio * top;
-    float left = -right;
-    Vector3f gaze = new Vector3f();
+    double dist = 1.f;
+    double top = dist * (double)Math.tan(this.fovy * Math.PI / 360.d);
+    double bottom = -top;
+    double right = aspect_ratio * top;
+    double left = -right;
+    Vector3d gaze = new Vector3d();
     gaze.sub(this.at, this.eye);
 
     center = this.eye;
-    Vector3f W = gaze;
+    Vector3d W = gaze;
     W.negate();
     W.normalize();
-    Vector3f V = this.up;
-    Vector3f U = new Vector3f();
+    Vector3d V = this.up;
+    Vector3d U = new Vector3d();
     U.cross(V, W);
     U.normalize();
     V.cross(W, U);
 
-    corner = new Vector3f();
+    corner = new Vector3d();
     corner.scaleAdd(left, U, center);
     corner.scaleAdd(bottom, V, corner);
     corner.scaleAdd(-dist, W, corner);
 
-    across = new Vector3f(U);
+    across = new Vector3d(U);
     across.scale(right-left);
 
-    this.up = new Vector3f(V);
+    this.up = new Vector3d(V);
     this.up.scale(top - bottom);
   }
 
-  Ray getCameraRay(float x, float y)
+  Ray getCameraRay(double x, double y)
   {
 		/* getCameraRay function
 		 * (x,y) is a normalized image coordinate, where
 		 * both of them vary between [0,1] */
 
-    Vector3f direction = new Vector3f();
+    Vector3d direction = new Vector3d();
     direction.scaleAdd(x, across, corner);
     direction.scaleAdd(y, up, direction);
     direction.sub(center);
