+++ src/edu/umass/cs390cg/atmosphere/geom/Ray.java
@@ -1,28 +1,28 @@
 package edu.umass.cs390cg.atmosphere.geom;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 public class Ray {
-  public Vector3f o,d;
+  public Vector3d o,d;
 
   public Ray() {
-    this.o = new Vector3f(0, 0, 0);
-    this.d = new Vector3f(0, -1, 0);
+    this.o = new Vector3d(0, 0, 0);
+    this.d = new Vector3d(0, -1, 0);
   }
 
   public Ray(Ray ray) {
-    this.o = new Vector3f(ray.o);
-    this.d = new Vector3f(ray.d);
+    this.o = new Vector3d(ray.o);
+    this.d = new Vector3d(ray.d);
   }
 
-  public Ray(Vector3f o, Vector3f d) {
-    this.o = new Vector3f(o);
-    this.d = new Vector3f(d);
+  public Ray(Vector3d o, Vector3d d) {
+    this.o = new Vector3d(o);
+    this.d = new Vector3d(d);
     this.d.normalize();
   }
 
-  public Vector3f pointAt(float t) {
-    Vector3f point = new Vector3f();
+  public Vector3d pointAt(double t) {
+    Vector3d point = new Vector3d();
     point.scaleAdd(t, d, o);
     return point;
   }
